requirements.txt
tensorflow-macos==2.13.0
larq==0.13.0
numpy==1.24.3
pandas==2.3.0
streamlit==1.45.1
matplotlib==3.10.3
scipy==1.10.1
pillow==11.2.1
h5py==3.14.0
Keras-Preprocessing==1.1.2